pyproject.toml
[project]
name = "tim-class-pass"
version = "0.1.0"
description = "A Python project for Tim Class Pass"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0.0",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.66.0",
    "openai>=2.26.0",
    "pillow>=10.0.0",
    "scikit-image>=0.22.0",
    "pendulum>=3.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
    "yt-dlp>=2026.3.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
]

[project.scripts]
tim-class-pass = "tim_class_pass.main:main"
tim-class-pass-usage = "tim_class_pass.usage_report:main"
extract-video-audio = "helpers.utils.video_audio_cli:main"
adjudication-export = "pipeline.adjudication.export_cli:main"
lesson-contract = "pipeline.contracts.cli:main"
corpus-cli = "pipeline.corpus.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/tim_class_pass", "src/helpers", "src/pipeline", "src/ml", "src/ui"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "live_provider: mark tests that call live LLM providers (run with -m live_provider); skip when API keys unset",
    "integration: slow tests that read or regenerate Lesson 2 pipeline artifacts",
    "e2e: browser end-to-end UI tests",
]
# src/ layout: packages are importable via src/ on the path (no editable install required)
pythonpath = ["src"]
testpaths = ["tests"]
# Frozen audit snapshots live under audit/; never collect tests from there
norecursedirs = ["audit", ".venv", "var", "dist"]
# importlib mode ensures pythonpath is applied before import even for __init__.py-based sub-packages
addopts = "--import-mode=importlib"

[dependency-groups]
dev = [
    "jsonschema>=4.23.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "pytest-playwright>=0.7.2",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*__pycache__*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true
